pyproject.toml
[tool.poetry]
name = "autodbdoc"
version = "0.1.0"
description = "A tool to generate Word documents from Oracle database metadata"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cx-Oracle = "^8.3.0"
python-docx = "^1.0.1"
python-dotenv = "^1.0.0"
flask = "^3.0.2"
flask-wtf = "^1.2.1"
flask-apscheduler = "^1.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 